Refactor: swagger 엔트리 포인트에서 분리
